perf(ui): extract ChatInput to fix typing lag in long chats
The `text` ref lived in TopicChat alongside `messages`, so every
keystroke triggered a full re-render of the entire message list.
Moving the textarea, file handling, compose overlay, and draft
persistence into a new ChatInput child component isolates keystroke
state to that component's render scope — message history is no longer
touched on input.

Also debounces localStorage draft saves (500 ms) instead of writing
on every keystroke.
